+++ src/main/java/com/gdsc_knu/official_homepage/annotation/TokenMemberResolver.java
@@ -1,7 +1,7 @@
 package com.gdsc_knu.official_homepage.annotation;
 
-import com.gdsc_knu.official_homepage.authentication.JwtTokenValidator;
-import com.gdsc_knu.official_homepage.authentication.redis.JwtMemberDetail;
+import com.gdsc_knu.official_homepage.authentication.jwt.JwtTokenValidator;
+import com.gdsc_knu.official_homepage.authentication.jwt.JwtMemberDetail;
 import com.gdsc_knu.official_homepage.entity.Member;
 import com.gdsc_knu.official_homepage.exception.CustomException;
 import com.gdsc_knu.official_homepage.exception.ErrorCode;
